app/requirements.txt
python-dotenv
prometheus-api-client
gevent
locust
numpy
pandas
requests
matplotlib
seaborn
docker
pyyaml
kubernetes
scikit-criteria
joblib
sklearn
extrap
scikit-learn
PySimpleGUI